Split AndroidBackend into multiple files
