+++ Sources/AndroidBackend/AndroidBackend.swift
@@ -347,7 +347,7 @@ public final class AndroidBackend: BackendFeatures.BaseStubs {
     }
 
     /// Converts a Swift String to a Java CharSequence.
-    private func charSequence(from string: String) -> CharSequence {
+    func charSequence(from string: String) -> CharSequence {
         let jstring = JavaString(string, environment: Self.env)
         return jstring.as(CharSequence.self)!
     }
@@ -446,514 +446,5 @@ public final class AndroidBackend: BackendFeatures.BaseStubs {
         let width = widget.getMeasuredWidth()
         let height = widget.getMeasuredHeight()
         return SIMD2(Int(width), Int(height))
-    }
-}
-
-extension AndroidBackend: BackendFeatures.WebViews {
-    public func createWebView() -> Widget {
-        CustomWebView(Self.activity, environment: Self.env).as(AndroidKit.View.self)!
-    }
-    
-    public func updateWebView(
-        _ webView: Widget,
-        environment: EnvironmentValues,
-        onNavigate: @escaping (URL) -> Void
-    ) {
-        let webView = webView.as(CustomWebView.self)!
-        webView.setOnNavigate(SwiftAction(environment: Self.env) {
-            if let javaString = webView.getLoadingUrl(),
-               let url = URL(string: javaString.toString()) {
-                onNavigate(url)
-            }
-        })
-    }
-    
-    public func navigateWebView(_ webView: Widget, to url: URL) {
-        let webView = webView.as(CustomWebView.self)!
-        webView.loadUrl(url.absoluteString)
-    }
-}
-
-extension AndroidBackend: BackendFeatures.Pickers {
-    public var supportedPickerStyles: [BackendPickerStyle] {
-        [.menu, .radioGroup, .wheel]
-    }
-
-    public func createPicker(style: BackendPickerStyle) -> Widget {
-        switch style {
-            case .radioGroup:
-                return CustomRadioGroup(
-                    Self.activity,
-                    environment: Self.env
-                ).as(AndroidKit.View.self)!
-            case .menu:
-                return CustomSpinner(
-                    Self.activity,
-                    environment: Self.env
-                ).as(AndroidKit.View.self)!
-            case .wheel:
-                return CustomNumberPicker(
-                    Self.activity,
-                    environment: Self.env
-                ).as(AndroidKit.View.self)!
-            default:
-                // TODO(bbrk24): Implement .segmented using MaterialButtonToggleGroup
-                fatalError("Unsupported picker style \(style)")
-        }
-    }
-    
-    public func updatePicker(
-        _ picker: Widget,
-        options: [String],
-        environment: EnvironmentValues,
-        onChange: @escaping (Int?) -> Void
-    ) {
-        if let picker = picker.as(CustomRadioGroup.self) {
-            let action = SwiftAction(environment: Self.env) {
-                let selectedOption = picker.getSelectedOption()
-                onChange(selectedOption < 0 ? nil : Int(selectedOption))
-            }
-            let textStyle = getTextStyle(from: environment)
-            picker.update(
-                action,
-                options,
-                environment.isEnabled,
-                color: textStyle.color,
-                fontSize: textStyle.fontSizePixels,
-                lineHeight: textStyle.lineHeightPixels,
-                textStyle.typeface
-            )
-        } else if let picker = picker.as(CustomSpinner.self) {
-            let action = SwiftAction(environment: Self.env) {
-                let selectedOption = picker.getSelectedItemPosition()
-                let invalidPosition: Int32 = try! JavaClass<AndroidKit.AdapterView>().INVALID_POSITION
-                
-                onChange(selectedOption == invalidPosition ? nil : Int(selectedOption))
-            }
-            picker.update(action, options, environment.isEnabled)
-        } else if let picker = picker.as(CustomNumberPicker.self) {
-            let action = SwiftAction(environment: Self.env) {
-                let selectedOption = picker.getValue()
-                onChange(selectedOption == 0 ? nil : Int(selectedOption - 1))
-            }
-            picker.update(action, options, environment.isEnabled)
-        } else {
-            fatalError("Unexpected picker class")
-        }
-    }
-    
-    public func setSelectedOption(ofPicker picker: Widget, to selectedOption: Int?) {
-        if let picker = picker.as(CustomRadioGroup.self) {
-            picker.selectOption(Int32(selectedOption ?? -1))
-        } else if let picker = picker.as(CustomSpinner.self) {
-            if let selectedOption {
-                picker.selectOption(Int32(selectedOption))
-            } else {
-                let invalidPosition: Int32 = try! JavaClass<AndroidKit.AdapterView>().INVALID_POSITION
-                
-                picker.selectOption(invalidPosition)
-            }
-        } else if let picker = picker.as(AndroidKit.NumberPicker.self) {
-            if let selectedOption {
-                picker.setValue(Int32(selectedOption + 1))
-            } else {
-                picker.setValue(0)
-            }
-        } else {
-            fatalError("Unexpected picker class")
-        }
-    }
-}
-
-extension AndroidBackend: BackendFeatures.Alerts {
-    public typealias Alert = AlertFragment
-    
-    public func createAlert() -> AlertFragment {
-        AlertFragment(environment: Self.env)
-    }
-    
-    public func updateAlert(
-        _ alert: AlertFragment,
-        title: String,
-        actionLabels: [String],
-        environment: EnvironmentValues
-    ) {
-        alert.update(title, actionLabels)
-    }
-    
-    public func showAlert(
-        _ alert: AlertFragment,
-        window: Window?,
-        responseHandler handleResponse: @escaping (Int) -> Void
-    ) {
-        let action = SwiftAction(environment: Self.env) {
-            let index = alert.getButtonIndex()
-            handleResponse(Int(index))
-        }
-        
-        alert.setAction(action)
-        alert.show(Self.activity)
-    }
-    
-    public func dismissAlert(_ alert: AlertFragment, window: Window?) {
-        alert.dismiss()
-    }
-}
-      
-extension AndroidBackend: BackendFeatures.ToggleButtons, BackendFeatures.Checkboxes, BackendFeatures.Switches {
-    public var requiresToggleSwitchSpacer: Bool { false }
-
-    public func createToggle() -> Widget {
-        AndroidKit.ToggleButton(
-            Self.activity,
-            environment: Self.env
-        )
-    }
-
-    public func createCheckbox() -> Widget {
-        AndroidKit.CheckBox(
-            Self.activity,
-            environment: Self.env
-        )
-    }
-
-    public func createSwitch() -> Widget {
-        AndroidKit.Switch(
-            Self.activity,
-            environment: Self.env
-        )
-    }
-
-    private func updateCompoundButton(
-        _ button: AndroidKit.CompoundButton,
-        environment: EnvironmentValues,
-        onChange: @escaping (Bool) -> Void
-    ) {
-        button.setEnabled(environment.isEnabled)
-
-        let action = SwiftAction(environment: Self.env) {
-            let checked = button.isChecked()
-            onChange(checked)
-        }
-        let listener = CustomOnCheckedChangeListener(action, environment: Self.env)
-
-        button.setOnCheckedChangeListener(
-            listener.as(AndroidKit.CompoundButton.OnCheckedChangeListener.self)!
-        )
-    }
-
-    public func updateToggle(
-        _ toggle: Widget,
-        label: String,
-        environment: EnvironmentValues,
-        onChange: @escaping (Bool) -> Void
-    ) {
-        let toggle = toggle.as(AndroidKit.ToggleButton.self)!
-        updateCompoundButton(toggle, environment: environment, onChange: onChange)
-
-        let charSequence = charSequence(from: label)
-        toggle.setTextOn(charSequence)
-        toggle.setTextOff(charSequence)
-
-        getTextStyle(from: environment).apply(to: toggle)
-    }
-
-    public func updateCheckbox(
-        _ checkboxWidget: Widget,
-        environment: EnvironmentValues,
-        onChange: @escaping (Bool) -> Void
-    ) {
-        let checkboxWidget = checkboxWidget.as(AndroidKit.CompoundButton.self)!
-        updateCompoundButton(checkboxWidget, environment: environment, onChange: onChange)
-    }
-
-    public func updateSwitch(
-        _ switchWidget: Widget,
-        environment: EnvironmentValues,
-        onChange: @escaping (Bool) -> Void
-    ) {
-        let switchWidget = switchWidget.as(AndroidKit.CompoundButton.self)!
-        updateCompoundButton(switchWidget, environment: environment, onChange: onChange)
-    }
-
-    public func setState(ofToggle toggle: Widget, to state: Bool) {
-        let toggle = toggle.as(AndroidKit.CompoundButton.self)!
-        toggle.setChecked(state)
-    }
-
-    public func setState(ofCheckbox checkboxWidget: Widget, to state: Bool) {
-        let checkboxWidget = checkboxWidget.as(AndroidKit.CompoundButton.self)!
-        checkboxWidget.setChecked(state)
-    }
-
-    public func setState(ofSwitch switchWidget: Widget, to state: Bool) {
-        let switchWidget = switchWidget.as(AndroidKit.CompoundButton.self)!
-        switchWidget.setChecked(state)
-    }
-}
-
-extension SwiftCrossUI.Color.Resolved {
-    func asColorInt() -> Int32 {
-        let alpha = UInt32(opacity * 255.0 + 0.5)
-        let red = UInt32(red * 255.0 + 0.5)
-        let green = UInt32(green * 255.0 + 0.5)
-        let blue = UInt32(blue * 255.0 + 0.5)
-
-        let combined = (alpha << 24) | (red << 16) | (green << 8) | blue
-
-        return Int32(bitPattern: combined)
-    }
-    
-    init(fromColorInt int: Int32) {
-        let uint = UInt32(bitPattern: int)
-
-        let alpha = Float(uint >> 24) / 255.0
-        let red = Float((uint & 0x00FF0000) >> 16) / 255.0
-        let green = Float((uint & 0x0000FF00) >> 8) / 255.0
-        let blue = Float(uint & 0x000000FF) / 255.0
-
-        self.init(
-            red: red,
-            green: green,
-            blue: blue,
-            opacity: alpha
-        )
-    }
-}
-
-extension AndroidBackend: BackendFeatures.Colors {
-    public func createColorableRectangle() -> Widget {
-        AndroidKit.View(Self.activity, environment: Self.env)
-    }
-    
-    public func setColor(
-        ofColorableRectangle widget: Widget,
-        to color: SwiftCrossUI.Color.Resolved
-    ) {
-        widget.setBackgroundColor(color.asColorInt())
-    }
-    
-    public func resolveAdaptiveColor(
-        _ adaptiveColor: SwiftCrossUI.Color.SystemAdaptive,
-        in environment: EnvironmentValues
-    ) -> SwiftCrossUI.Color.Resolved {
-        let Rcolor = try! JavaClass<AndroidKit.R.color>()
-
-        let resId: Int32? =
-            switch (adaptiveColor, environment.colorScheme) {
-                case (.blue, .light): Rcolor.holo_blue_dark
-                case (.blue, .dark): Rcolor.holo_blue_light
-                case (.gray, _): Rcolor.darker_gray
-                case (.green, .light): Rcolor.holo_green_dark
-                case (.green, .dark): Rcolor.holo_green_light
-                case (.orange, .light): Rcolor.holo_orange_dark
-                case (.orange, .dark): Rcolor.holo_orange_light
-                case (.red, .light): Rcolor.holo_red_dark
-                case (.red, .dark): Rcolor.holo_red_light
-                case (.purple, _): Rcolor.holo_purple
-                default: // brown, yellow
-                    nil
-            }
-        
-        guard let resId else {
-            return SwiftCrossUI.Color.defaultResolveAdaptiveColor(adaptiveColor, in: environment)
-        }
-        
-        let colorInt = Self.activity.getColor(resId)
-        
-        return SwiftCrossUI.Color.Resolved(fromColorInt: colorInt)
-    }
-}
-
-// TODO(bbrk24): When everything else uses dp, also scale stroke width and actions
-extension AndroidBackend: BackendFeatures.Paths {
-    public struct Path {
-        var path: AndroidGraphics.Path
-        var fillPaint: AndroidKit.Paint
-        var strokePaint: AndroidKit.Paint
-
-        init(env: JNIEnvironment?) {
-            path = AndroidGraphics.Path(environment: env)
-
-            let styleClass = try! JavaClass<AndroidKit.Paint.Style>()
-
-            fillPaint = AndroidKit.Paint(environment: env)
-            fillPaint.setStyle(styleClass.FILL)
-
-            strokePaint = AndroidKit.Paint(environment: env)
-            strokePaint.setStyle(styleClass.STROKE)
-        }
-    }
-
-    public func createPathWidget() -> Widget {
-        PathView(Self.activity, environment: Self.env)
-            .as(AndroidKit.View.self)!
-    }
-
-    public func createPath() -> Path {
-        Path(env: Self.env)
-    }
-
-    public func updatePath(
-        _ path: Path,
-        _ source: SwiftCrossUI.Path,
-        bounds: SwiftCrossUI.Path.Rect,
-        pointsChanged: Bool,
-        environment: EnvironmentValues
-    ) {
-        apply(strokeStyle: source.strokeStyle, to: path)
-
-        let fillTypeClass = try! JavaClass<AndroidGraphics.Path.FillType>()
-
-        switch source.fillRule {
-            case .evenOdd:
-                path.path.setFillType(fillTypeClass.EVEN_ODD)
-            case .winding:
-                path.path.setFillType(fillTypeClass.WINDING)
-        }
-
-        if pointsChanged {
-            path.path.reset()
-            apply(actions: source.actions, to: path.path)
-        }
-    }
-
-    public func renderPath(
-        _ path: Path,
-        container: Widget,
-        strokeColor: SwiftCrossUI.Color.Resolved,
-        fillColor: SwiftCrossUI.Color.Resolved,
-        overrideStrokeStyle: StrokeStyle?
-    ) {
-        if let overrideStrokeStyle {
-            apply(strokeStyle: overrideStrokeStyle, to: path)
-        }
-
-        path.strokePaint.setColor(strokeColor.asColorInt())
-        path.fillPaint.setColor(fillColor.asColorInt())
-
-        container.as(PathView.self)!.set(
-            path: path.path,
-            fillPaint: path.fillPaint,
-            strokePaint: path.strokePaint
-        )
-    }
-
-    private func apply(strokeStyle: StrokeStyle, to path: Path) {
-        path.strokePaint.setStrokeWidth(Float(strokeStyle.width))
-
-        let capClass = try! JavaClass<AndroidKit.Paint.Cap>()
-        switch strokeStyle.cap {
-            case .butt:
-                path.strokePaint.setStrokeCap(capClass.BUTT)
-            case .round:
-                path.strokePaint.setStrokeCap(capClass.ROUND)
-            case .square:
-                path.strokePaint.setStrokeCap(capClass.SQUARE)
-        }
-
-        let joinClass = try! JavaClass<AndroidKit.Paint.Join>()
-        // Even though fillPaint doesn't render a line, the stroke join can still affect the shape
-        // of the fill.
-        switch strokeStyle.join {
-            case .miter(let limit):
-                path.strokePaint.setStrokeJoin(joinClass.MITER)
-                path.strokePaint.setStrokeMiter(Float(limit))
-
-                path.fillPaint.setStrokeJoin(joinClass.MITER)
-                path.fillPaint.setStrokeMiter(Float(limit))
-            case .bevel:
-                path.strokePaint.setStrokeJoin(joinClass.BEVEL)
-                path.fillPaint.setStrokeJoin(joinClass.BEVEL)
-            case .round:
-                path.strokePaint.setStrokeJoin(joinClass.ROUND)
-                path.fillPaint.setStrokeJoin(joinClass.ROUND)
-        }
-    }
-
-    private func apply(actions: [SwiftCrossUI.Path.Action], to path: AndroidGraphics.Path) {
-        lazy var directionClass = try! JavaClass<AndroidGraphics.Path.Direction>()
-
-        for action in actions {
-            switch action {
-                case .moveTo(let point):
-                    path.moveTo(Float(point.x), Float(point.y))
-                case .lineTo(let point):
-                    path.lineTo(Float(point.x), Float(point.y))
-                case .quadCurve(let control, let end):
-                    path.quadTo(
-                        Float(control.x),
-                        Float(control.y),
-                        Float(end.x),
-                        Float(end.y)
-                    )
-                case .cubicCurve(let control1, let control2, let end):
-                    path.cubicTo(
-                        Float(control1.x),
-                        Float(control1.y),
-                        Float(control2.x),
-                        Float(control2.y),
-                        Float(end.x),
-                        Float(end.y)
-                    )
-                case .rectangle(let rect):
-                    path.addRect(
-                        Float(rect.x),
-                        Float(rect.y),
-                        Float(rect.maxX),
-                        Float(rect.maxY),
-                        directionClass.CW
-                    )
-                case .circle(let center, let radius):
-                    path.addCircle(
-                        Float(center.x),
-                        Float(center.y),
-                        Float(radius),
-                        directionClass.CW
-                    )
-                case .arc(let center, let radius, let startAngle, let endAngle, let clockwise):
-                    var sweepAngle: Double
-                    if clockwise {
-                        if startAngle < endAngle {
-                            sweepAngle = endAngle - startAngle
-                        } else {
-                            sweepAngle = 2 * .pi + endAngle - startAngle
-                        }
-                    } else {
-                        if startAngle < endAngle {
-                            sweepAngle = endAngle - startAngle - 2 * .pi
-                        } else {
-                            sweepAngle = endAngle - startAngle
-                        }
-                    }
-
-                    path.addArc(
-                        Float(center.x - radius),
-                        Float(center.y - radius),
-                        Float(center.x + radius),
-                        Float(center.y + radius),
-                        Float(startAngle * 180 / .pi),
-                        Float(sweepAngle * 180 / .pi)
-                    )
-                case .transform(let transform):
-                    let matrix = AndroidKit.Matrix(environment: Self.env)
-                    matrix.setValues([
-                        Float(transform.linearTransform.x),
-                        Float(transform.linearTransform.y),
-                        Float(transform.translation.x),
-                        Float(transform.linearTransform.z),
-                        Float(transform.linearTransform.w),
-                        Float(transform.translation.y),
-                        0,
-                        0,
-                        1
-                    ])
-                    path.transform(matrix)
-                case .subpath(let actions):
-                    let subPath = AndroidGraphics.Path(environment: Self.env)
-                    apply(actions: actions, to: subPath)
-                    path.addPath(subPath)
-            }
-        }
     }
 }
